We would like to ask you to implement a converter between usual time representation and Berlin one. To do that task you have to realize what is the Berlin Clock .
The Berlin Clock is based on the principles of quantity didactics and
with this its the first of its kind. The time is calculated by adding
the lit rectangular lamps. The top lamp is a pump which is blinking on/off every two seconds.
In the upper line of red lamps every lamp represents 5 hours.
In the lower line of red lamps every lamp represents 1 hour.
So if in the first line 2 lamps are lit and in the second line 3 lamps its 5+5+3=13h
or 1 p.m. In the third line with tall lamps every lamp represents 5 minutes.
There are 11 lamps, the 3rd, 6th, and 9th are red indicating the first quarter, half,
and the last quarter of the hour. In the last line with yellow lamps every lamp represents
1 minute.

More information information can be found there.
https://en.wikipedia.org/wiki/Mengenlehreuhr